Movement as should have always been. @Simon

+++ src/Cameras/Camera.cpp
@@ -100,9 +100,9 @@ void Camera::ProcessMouseScroll(GLfloat yoffset) {
 void Camera::updateCameraVectors() {
     // Calculate the new Front vector
     glm::vec3 front;
-    front.x = cos(glm::radians(this->Yaw)) * cos(glm::radians(this->Pitch));
-    front.y = sin(glm::radians(this->Pitch));
-    front.z = sin(glm::radians(this->Yaw)) * cos(glm::radians(this->Pitch));
+    front.x = cosf(glm::radians(this->Yaw)) * cosf(glm::radians(this->Pitch));
+    front.y = sinf(glm::radians(this->Pitch));
+    front.z = sinf(glm::radians(this->Yaw)) * cosf(glm::radians(this->Pitch));
     this->Front = glm::normalize(front);
     // Also re-calculate the Right and Up vector
     this->Right = glm::normalize(glm::cross(this->Front, this->WorldUp));  // Normalize the vectors, because their length gets closer to 0 the more you look up or down which results in slower movement.
